+++ tests/integration-tests-isolated/pom.xml
@@ -23,9 +23,10 @@
       <version>2.29.0-SNAPSHOT</version>
    </parent>
 
-   <artifactId>integration-tests</artifactId>
+   <artifactId>integration-tests-isolated</artifactId>
    <packaging>jar</packaging>
-   <name>ActiveMQ Artemis Integration Tests</name>
+   <name>ActiveMQ Artemis Isolated Integration Tests</name>
+   <description>This project is pretty much a copy of integration-tests however the tests here will run with fork=always, isolating each test on its own VM. </description>
 
    <properties>
       <activemq.basedir>${project.basedir}/../..</activemq.basedir>
@@ -269,11 +270,28 @@
          <artifactId>postgresql</artifactId>
          <scope>test</scope>
       </dependency>
-
       <dependency>
-          <groupId>org.apache.commons</groupId>
-          <artifactId>commons-lang3</artifactId>
-          <scope>test</scope>
+         <groupId>org.apache.directory.server</groupId>
+         <artifactId>apacheds-test-framework</artifactId>
+         <version>${directory-version}</version>
+         <scope>test</scope>
+         <exclusions>
+            <exclusion>
+               <groupId>org.apache.directory.api</groupId>
+               <artifactId>api-ldap-schema-data</artifactId>
+            </exclusion>
+         </exclusions>
+      </dependency>
+      <dependency>
+         <groupId>org.apache.directory.server</groupId>
+         <artifactId>apacheds-server-annotations</artifactId>
+         <version>${directory-version}</version>
+         <scope>test</scope>
+      </dependency>
+      <dependency>
+         <groupId>org.apache.commons</groupId>
+         <artifactId>commons-lang3</artifactId>
+         <scope>test</scope>
       </dependency>
 
       <dependency>
@@ -291,6 +309,18 @@
       <dependency>
          <groupId>org.apache.hadoop</groupId>
          <artifactId>hadoop-minikdc</artifactId>
+         <scope>test</scope>
+      </dependency>
+      <dependency>
+         <groupId>org.apache.directory.server</groupId>
+         <artifactId>apacheds-interceptor-kerberos</artifactId>
+         <version>${directory-version}</version>
+         <scope>test</scope>
+      </dependency>
+      <dependency>
+         <groupId>org.apache.directory.jdbm</groupId>
+         <artifactId>apacheds-jdbm2</artifactId>
+         <version>${directory-jdbm2-version}</version>
          <scope>test</scope>
       </dependency>
       <dependency>
@@ -412,7 +442,8 @@
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-surefire-plugin</artifactId>
             <configuration>
-               <skipTests>${skipIntegrationTests}</skipTests>
+               <forkMode>always</forkMode>
+               <skipTests>${skipIsolatedIntegrationTests}</skipTests>
                <argLine>-Djgroups.bind_addr=::1 ${activemq-surefire-argline} ${its-surefire-extra-args} -Dorg.apache.activemq.SERIALIZABLE_PACKAGES="java.lang,javax.security,java.util,org.apache.activemq,org.fusesource.hawtbuf"</argLine>
             </configuration>
          </plugin>
@@ -420,31 +451,9 @@
    </build>
    <profiles>
       <profile>
-         <!-- Quick subset of tests, used on PR checks -->
-         <id>fast-tests</id>
-         <build>
-            <pluginManagement>
-               <plugins>
-                  <plugin>
-                     <groupId>org.apache.maven.plugins</groupId>
-                     <artifactId>maven-surefire-plugin</artifactId>
-                     <configuration>
-                        <includes combine.children="override">
-                           <include>**/integration/security/*Test.java</include>
-                           <include>**/integration/ssl/*Test.java</include>
-                           <include>**/integration/amqp/JMSSaslExternalTest.java</include>
-                           <include>**/integration/amqp/connect/AMQPConnectSaslTest.java</include>
-                        </includes>
-                     </configuration>
-                  </plugin>
-               </plugins>
-            </pluginManagement>
-         </build>
-      </profile>
-      <profile>
          <id>jdk16on</id>
-            <activation>
-         <jdk>[16,)</jdk>
+         <activation>
+            <jdk>[16,)</jdk>
          </activation>
          <properties>
             <its-surefire-extra-args>--add-exports java.security.jgss/sun.security.krb5=ALL-UNNAMED</its-surefire-extra-args>
