+++ tests/integration-tests-isolated/src/test/resources/AMQauth.ldif
