+++ tests/integration-tests-isolated/src/test/resources/AMQauth2.ldif
