+++ tests/integration-tests-isolated/src/test/resources/AMQauth3.ldif
