+++ tests/integration-tests-isolated/src/test/resources/SaslKrb5LDAPSecurityTest.ldif
