+++ tests/pom.xml
@@ -130,6 +130,7 @@
       <module>timing-tests</module>
       <module>jms-tests</module>
       <module>integration-tests</module>
+      <module>integration-tests-isolated</module>
       <module>karaf-client-integration-tests</module>
       <module>karaf-server-integration-tests</module>
       <module>compatibility-tests</module>
